lost_and_found/requirements.txt
streamlit
torch
torchvision
ftfy
regex
tqdm
pillow
scikit-learn
git+https://github.com/openai/CLIP.git
